Fix CardView*
*A QuizCard inflated into the collapsed state cannot animate into the
expanded state because it never had time to measure itself in the
expanded state. This measurement requires waiting for layout to occur
on screen. To overcome this we can “simply” have quiz_card_choice.xml
have a fixed height based on some resource dimension. We use that
dimension multiplied by the number of quiz choices + inter-choice
padding to calculate the expanded view without having to rely on layout
giving it to us.