ClipCard update via CardAdapter#notifyCardChanged
After much digging and scratching a moment of clarity! Let the adapter
handle all data changes with UI consequence: that’s its job! A clean
distinction emerges between the role of Observers (managing changes to
the state of the model) and our CardAdapter (responding to the subset
of those changes that should affect the UI). A possible improvement
could be having the CardAdapter actually observe all Cards passed to it
via the add methods and its Constructor. Then we’d lose the burden of
calling the CardAdapter#addCard/changeCard manually.