ClipCard clip importer
Uses the Storage Access Framework for Android 4.4 up, and regular
Intent.ACTION_GET_CONTENT below.

NOTE: The Storage Access Framework only reveals files by its
ContentProvider Uri, and thus the data is only accessible as a Stream.
This is actually more convenient, except that we need a Context to
access the ContentResolver and so cannot override
MediaFile#getThumbnail as it has no Context. Thus for now ClipCardView
and ReviewCardView load the pre-generated thumbnail from the system
ContentProvider if they recognize the content://xxx path type in
MediaFile#getPath()