Revamp thumbnail management
Thumbnail generation is now asynchronous and the storage locations for
thumbnails and audio recordings have been made consistent.

In the external storage, we now have:

— / Liger
    |- / audio (.mp4 audio recordings)
    |- / thumbs (audio, photo, video, thumbnails)

Remember that video and photos are stored in the system location
(likely /sdcard/DCIM). However we now no longer store thumbnails in
that directory, which polluted the system gallery.